KKUCAssistant/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Hoang Le"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = "^0.3.2"
pydantic = "^2.9.2"
langchain-core = "^0.3.17"
langchain-openai = "^0.2.8"
langgraph = "^0.2.46"
python-dotenv = "^1.0.1"
firecrawl-py = "^1.0.0"
openai = "^1.0.0"
cohere = "^5.11.0"
weaviate-client = "^4.9.0"
rank-bm25 = "^0.2.2"
langsmith = "^0.1.0"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
